+++ src/FsCodec/FsCodec.fs
@@ -1,7 +1,7 @@
 namespace FsCodec
 
-/// Common form for either a Domain Event or an Unfolded Event
-type IEvent<'Format> =
+/// Common form for either a Domain Event or an Unfolded Event, without any context regarding its place in the timeline of events
+type IEventData<'Format> =
     /// The Event Type, used to drive deserialization
     abstract member EventType : string
     /// Event body, as UTF-8 encoded json ready to be injected into the Store
@@ -15,43 +15,43 @@ type IEvent<'Format> =
     /// </remarks>
     abstract member Timestamp : System.DateTimeOffset
 
-/// Represents a Domain Event or Unfold, together with it's Index in the event sequence
-type IIndexedEvent<'Format> =
-    inherit IEvent<'Format>
-    /// The index into the event sequence of this event
+/// Represents a Domain Event or Unfold, together with it's 0-based <c>Index</c> in the event sequence
+type ITimelineEvent<'Format> =
+    inherit IEventData<'Format>
+    /// The 0-based index into the event sequence of this Event
     abstract member Index : int64
-    /// Indicates this is not a Domain Event, but actually an Unfolded Event based on the state inferred from the events up to `Index`
+    /// Indicates this is not a true Domain Event, but actually an Unfolded Event based on the State inferred from the Events up to and including that at <c>Index</c>
     abstract member IsUnfold : bool
 
 /// Defines a contract interpreter for a Discriminated Union representing a set of events borne by a stream
 type IUnionEncoder<'Union, 'Format> =
     /// Encodes a union instance into a decoded representation
-    abstract Encode : value: 'Union -> IEvent<'Format>
-    /// Decodes a formatted representation into a union instance. Does not throw exception on format mismatches
-    abstract TryDecode : encoded: IIndexedEvent<'Format> -> 'Union option
+    abstract Encode : value: 'Union -> IEventData<'Format>
+    /// Decodes a formatted representation into a <c>'Union<c> instance. Does not throw exception on format mismatches
+    abstract TryDecode : encoded: ITimelineEvent<'Format> -> 'Union option
 
 namespace FsCodec.Core
 
 open System
 
-/// An Event about to be written, see IEvent for further information
+/// An Event about to be written, see <c>IEventData<c> for further information
 [<NoComparison; NoEquality>]
 type EventData<'Format> private (eventType, data, meta, timestamp) =
     static member Create(eventType, data, ?meta, ?timestamp) =
         EventData(eventType, data, defaultArg meta Unchecked.defaultof<_>, match timestamp with Some ts -> ts | None -> DateTimeOffset.UtcNow)
-    interface FsCodec.IEvent<'Format> with
+    interface FsCodec.IEventData<'Format> with
         member __.EventType = eventType
         member __.Data = data
         member __.Meta = meta
         member __.Timestamp = timestamp
 
-/// An Event that's been read from a Store
+/// An Event or Unfold that's been read from a Store and hence has a defined <c>Index</c> on the Event Timeline
 [<NoComparison; NoEquality>]
-type IndexedEventData<'Format> private (index, isUnfold, eventType, data, meta, timestamp) =
+type TimelineEvent<'Format> private (index, isUnfold, eventType, data, meta, timestamp) =
     static member Create(index, eventType, data, ?meta, ?timestamp, ?isUnfold) =
         let isUnfold, meta = defaultArg isUnfold false, defaultArg meta Unchecked.defaultof<_>
-        IndexedEventData(index, isUnfold, eventType, data, meta, match timestamp with Some ts -> ts | None -> DateTimeOffset.UtcNow)
-    interface FsCodec.IIndexedEvent<'Format> with
+        TimelineEvent(index, isUnfold, eventType, data, meta, match timestamp with Some ts -> ts | None -> DateTimeOffset.UtcNow)
+    interface FsCodec.ITimelineEvent<'Format> with
         member __.Index = index
         member __.IsUnfold = isUnfold
         member __.EventType = eventType